Enable recycling of third party preference widgets
Until now we have been recycling only framework preference widgets since
there were no guarantees for third party preferences to support recycling.
This let to broken animations for preference widgets that are outside of
the framework.

This change makes recycling to be used always and in case the developers
need to turn it off they can explicitely use a new attribute that is
being added to the Preference called "recycleEnabled" and set it to false.

Bug: b/34334451
Test: Test are part of the same topic.

Change-Id: I324087841e1edddbf0d3eaad00b5895a196acff6

Former-commit-id: 4186b344d829e155feebf6797bd41576fccb5f16