Add drag down to dismiss pip as a dismiss option
If enabled via flag, when the user drags the PIP past
a threshold at the bottom of the screen, the PIP will
be dismissed. If the PIP is sitting near the bottom
edge and the user flings downwards the PIP will be
dismissed.

Unifies logic to identify for “fling towards edge” for
minimize and dismiss gesture — adds additional logic
to check that the start of the fling gesture was near
that edge. This stops you from accidentally minimizing
or dismissing on a “long fling”.

Test: manual - enable flag, and drag pip to bottom of
      screen to dismiss
Bug: 35358628
Change-Id: Iddd66ed2720964d431cddea6173767eb5565b991

Former-commit-id: 57d2255cc983b614be61cab86ea8d5cafc03fdec