NativeActivity JNI: Retain VM reference to AssetManager
AssetManagers get destroyed when in MultiWindow or when loading
WebView. Retain an reference to the AssetManager given to
Native Code.

This AssetManager does not need to be updated since the changes
happening to the AssetManager involve non-application assets and
configuration changes, both which don't affect the native AAssetManager
which can only load files in assets/ which are unaffected by
configuration changes.

Test: none
Change-Id: Iedf30212fb85a8e6afec0645d689288f58fd77f1

Former-commit-id: 5c690d55dd2d0286db1ba3331a64769c07636ff8