Added ETWS primary message default message support
ETWS primary message does not contain messagy body. We used
to use hardcoded "ETWS" as the body, which is not easy for
end users to understand what happened. Added the built-in
default messages provided by Japanese government guideline
for earthquake, Tsunami, test, and other channels support to
enhance the user experience.

Test: manual
bug: 33595007
Change-Id: I53659298d8ce3b3ba5c751c14c1b36e76d6dc562

Former-commit-id: e3070ede55652bb00618eadfac9daee64320569f