Do not defer wm display logic when called from am.
Currently we process any remove/add/change logic from the activity
manager to the window manager through a handler message. This opens
the possibility of the activity manager requesting display state
from a window manager that is a step behind.

This change modifies the call behavior so that these changes are done
synchronously between the two managers.

Change-Id: Icd19f37c62bfb7aad10fab2e1b21e41ac75052b5
Fixes: 35741135
Test: cts/hostsidetests/services/activityandwindowmanager/util/run-test CtsServicesHostTestCases android.server.cts.ActivityManagerDisplayTests#testCreateMultipleVirtualDisplaysH

Former-commit-id: af3a400ead86f29ca355613dbe3a13b6fc3243d3