Add a default Surface to the VR Virtual Display.
Uses ImageReader as a do-nothing Surface to ensure the Virtual Display
gets passed through to SurfaceFlinger.

Without a Surface, the Virtual Display never makes it through the
android system to SurfaceFlinger. This change is needed to get the
Virtual Display recognized in the Flinger layers.

Test: 'setprop vr_virtualdisplay true' to enable virtual displays.
'setprop vr_debug_vd_receiver true' and shell restart for the test app.

Bug: 36491910
Change-Id: I2bf2cb4035dcf04484b7fb0bec61671864069335

Former-commit-id: 975c00d8ffeeb72a05bc229e0e1234e771c4a824