TimePickerDialog OTG keyboard entry discarded
A user can manually enter time value using a keyboard.
NumberPicker then evaluates its value when focus is changed.
Currently when the dialog is closed by pressing OK, the value
from the focused NumberPicker is not taken into consideration.

To ensure retrieval of the correct value when closing the
dialog, the focus must be removed from the NumberPicker as
is done in DatePickerDialog.

Bug: 65664546
Test: Manual
      1. Connect physical keyboard via OTG connector
      2. Enter Settings and set time manually
      3. Select keyboard entry
      4. Change time and tap OK
      The entered time should be applied
Change-Id: I8a77cb3aaa54acb228ec1ce0e9385e2eb5e88e9b

Former-commit-id: 10864b0fcbe5e4565bebf0690fc0c58153b4bea7