Remove tests from SettingsLibShell
The tests and the robolectric libraries are already in
SettingsLibRoboTests, and shouldn't be in the package being tested.
Including robolectric in the package causes it to go through
proguard, which complains about missing dependencies.

Also remove the unnecessary libraries that are already included
through frameworks/base/packages/SettingsLib/common.mk.

Test: m -j ANDROID_COMPILE_WITH_JACK=false RunSettingsLibRoboTests
Change-Id: I11f007c4321bb9b147f62010eae046f75d1a52e3

Former-commit-id: 3c17d776108680972bc50b21ee2e586d1c154d91