Update xattrs when creating cache directories.
Typically installd will create cache directories for an app, and
store those inode numbers in the parent directory.  However, when an
app deletes and recreates a cache directory, the (now invalid) inode
number is left floating around, and we could end up pointing at a
useful non-cache directory.

Fix this by updating the xattr when we create the cache directory
out in userspace.

Test: builds, boots
Bug: 34185870
Change-Id: I71526dd73230c938aceabd2c0a8e89d44cc1f169

Former-commit-id: 83aacdee407a14c15e5ac511b288df6bef1d63cb