Listen to USER_STARTED instead of USER_ADDED from WebViewUpdateService.
When adding a restricted user, 'hidden' applications are uninstalled for
the user after creation of the user, but before the user is 'started'.

Because WebViewUpdateService only listens to USER_ADDED and USER_REMOVED
intents it wouldn't realize that Chrome had been uninstalled for a new
restricted user (at the time of user-creation Chrome would be
installed). Instead listen to user-started/stopped intents to ensure the
restricted-user case is handled correctly.

Bug: 35813523
Test: Add a restricted user (on a Pixel C), with Chrome disabled. Start
an app using WebView (and ensure WebView doesn't crash).

Change-Id: I0c5d0a543b4bc80d269e586444b80cd3baffa11c

Former-commit-id: a43f144d4f0443146bae33359cc3222d2059db16