Drop REASON_NON_SYSTEM_LIBRARY.
Dexopt them with the same "mode" as the apk we are
going to dexopt. Because we treat them as shared, this
means they are only going to be fully compiled once an app
that has it in its <uses-library> gets compiled.

bug:33799337
Test: m -j32 && boot
Change-Id: Ic93fbcf45be6b9cf31ad3a3eb498f9fbbefeb288
(cherry picked from commit 6613326d5a1f757778743321b4987e3077d11056 [formerly 90978b4159dd1394c0643ae8171d24668fbc27a4])

Former-commit-id: de0272b85c4ff1eba1c58a3cdfd53bd5eb9925ef