Reschedule jobs that crash partway through
We were properly re-running jobs if the app hung during their
execution, but outright crashes wound up with the scheduled job
being dropped.  That's bad; it can easily lead to broken monitoring
in the case of content-triggered jobs.

We now reschedule with backoff when this happens.  In addition, to
mitigate the impact of repeatedly crashing apps, we now enforce a
minimum backoff interval of 10 seconds for automatic reschedules.

Bug 36030229
Test: manual

Change-Id: Ib5da583d7901d1255066c48558b35186eb641e17

Former-commit-id: b2087f874c9f370b813df49a180609a3ef901a62