Hide API to listen to a11y service state
Hiding and adding the requested return value, so if we
later decide to unhide it we'll be up to snuff for API
review.

Bug: 35764675
Test: Verified that the two features that listen, a11y
volume and the accessibility button on the nav bar, work
properly. Also ran accessibility CTS.
Merged-In: I47b62f2b4fd1ea4a7e697d20a3bb2c3201f6bcce
Change-Id: I4df6164772eb01b4f988c201f46c788eb7124511

Former-commit-id: 4cab9308fc73967f8085325b237d30978a7ba6da