Fix saving target fragment when Fragment not yet created.
Bug 35928527

During optimized transactions, a fragment may be removed without
being created. That leaves the state of the fragment in INITIALIZING
and previously, that state wasn't ever saved. This CL allows a
fragment that is being removed to be brought up to the CREATED state
so that it can be saved during saveAllState().

Test: manual and Ie7207cc647312d38b377405bc5ec8721db757d2e

Change-Id: I649f1931745be43087ec3578e9195624e80821dc

Former-commit-id: 5daf06c1e96fcfd38bfc0e58471ec40e7bd838f1