Offer to "fixup" GIDs used for app data.
We recently started tracking cached app data using a per-app GID for
the "cache" and "code_cache" directories and their contents.  For
upgraded devices, we ideally want to "fixup" the GIDs of any existing
data while the device is still showing the boot animation, instead of
blocking the user when they unlock the device.

This fixup operation has the nice side-effect of priming the dentry
cache with all the directories that we'll be validating when unlocking
a user, so by always running fixup we speed up onBeforeUnlockUser()
from about ~540ms to ~220ms.

Test: /data/nativetest/installd_service_test/installd_service_test
Bug: 34201111, 35084485
Change-Id: If40244d0356beb11df25d3052097c3c5a90d010f

Former-commit-id: 1566233bfdb994cd41d21ffb2615af8d0e29166f