Consider the last inserted SIM for EmergencyAffordance
If India Sim is inserted and removed in Airplane mode
and reboot, The device is supposed to consider the
previous status based on the last inserted SIM as
there wont be any cell info update in Airplane mode.

Rootcause is when there are no active subscription
list right from the bootup, EmergencyAffordance global
variable is not updated with the settings based on
previous boot.

Fix is to update the EmergencyAffordance even when
there are no active subscriptions available from
bootup.

Test: manual, see repro steps on bug
Fixes: 36978415
Merged-In: I072cb635eb57c63dd94bf8f0a0b6222e22fdfbb7
Change-Id: I072cb635eb57c63dd94bf8f0a0b6222e22fdfbb7

Former-commit-id: 3d465761d7bc81a763827bf80ac054b67a931ab5