Force direction of phone number fields to (practically) LTR
Previously, phone number fields defaulted to locale direction, which
created problems for apps that were not aware of LTR requirements for
phone numbers.

Now, we look at the direction of the digits for the locale, and use
that to determine the direction of the edit field. (For practically
all major RTL locales, that direction is LTR.)

Test: Manual
Bug: 33643035
Change-Id: I17c70d8462bd403ea6866057971105f1f5772ba3

Former-commit-id: 380154724b513ff7649d93f03d67b0e481f0db02