Move inner classes of RefactoredBackupManagerService to their own files.
The move was done automatically separately for each class. For the
classes that are not static, instance of RBMS is passed in the
constructor.

To make this compilable all methods and fields of RBMS that are accessed
from the separated classes were made public. In later commits this will
be addressed.

Test: built, flashed, verified that bmgr backupnow works with RBMS
enabled.

Bug: 36850431

Change-Id: I2d9be8bb045d5c4d8106bc59630d3a278138b54f

Former-commit-id: f251e3509838e3fbc62ccdba9d4cfd0527f67acd