Allow notification expansion to be toggleable.
Android Auto does not want notificaitons that have actions to be
expandable. Rather, they should always be in their expanded state. The
only exception to this are grouped notifications, which should be
collasped by default. They also can still be expanded.

Modify the code to allow this to be changed. In addition, modify some of
the animation so that the dividers so not flash when the notification
group is expanding.

There is also a change to bring the header of the notification to the
front so that the child notifications do not overlap it. This is because
in Auto, the app name and icon are not displayed in the header, but the
expand button is still there.

Test: booted on phone and Android Auto headunit.
Bug: 33210494
Change-Id: I0a121a0eaee6159ccea9961721abe92603484a17

Former-commit-id: 6bf88a01d7811d9460e088e743882a02d6250b86