Don't end task in finishActivityLocked if task is being reused
In scenarios when the task is being cleared for reuse, endTask
is true in finishActivityLocked even though the task is not actually
going to be removed later on. On Arc++, this leads to issues
because the notifyTaskRemovalStarted notification is delivered
and cleanup work happens for a task that isn't actually removed.

Fix this by looking at TaskRecord.mReuseTask, which is set to
true when the task is cleared in order to be reused.

Cherry-picked from
I3c422f57a8cda55f7b0c08e183f4f1a15d3b7ddd

Bug: 37093330
Test: Launch a test activity that relaunches itself with
FLAG_ACTIVITY_NEW_TASK AND FLAG_ACTIVITY_CLEAR_TASK,  verify that
notifyTaskRemovalStarted call back is not sent.

Change-Id: Ifee23ad7104124e38b5ee5e6c520eb25a5846ef2

Former-commit-id: bdef5377759da4571b72cc560d8a43bc4eee1a00