Redesign of QSSecurityFooter
QSSecurityFooter is redesigned to display the new strings related to
Device Management, Network Logging, VPNs and CA Certificates. The dialog
will be changed in a separate CL.

As part of this CL, the network logging icon and the WiFi-Assist branding
is removed.

Details of the design can be found in the bug.

Bug: 34369166
Test: in class QSSecurityFooterTest, which is part of "runtest systemui"

Change-Id: I71cd57ff607fe00d3ee0aba2312143f728655c76

Former-commit-id: 774c6542cd429e33a8dcecd9c775d61f3675cdaf