Add getVisualVoicemailSettings()
The voicemail client in telephony is moved to the system dialer
in O. Some user data such as the enabled status of VVM or the
voicemail PIN scrambled by the client need to be migrated to the
system dialer for it to function seamlessly.

This CL adds getVisualVoicemailSettings() to let the system dialer
retrieve a opaque bundle to migrate the user data. The system dialer
and telephony service should agreed upon the structure and content
of the bundle.

Bug: 34093562
Test: N/A - pending dialer side changes to be implemented.
Change-Id: Ib7da801ada95451e55e02eda520894aa41316f73

Former-commit-id: 1210f41f59a131b50e4d54f308e7d1acbe76cbea