Window can only be fully drawn if app window is not relaunching
This fixes the glitch of having transitions be calculated incorrectly
when starting new activity when other windows still exist. For example
launcher3 has minus one window exist even if the main activity is
destroyed causing incorrect transition going home. This change only
allows the calculation to occur when the app window is not relaunching.

Test: manual - launch app, rotate screen, go home
Change-Id: I7e323486ee6e05ba4c8704832ca0f9d6e648ac90
Fixes: 36113180

Former-commit-id: 498c71dd9dbcf2b13e0beb923a027a302e2cf685