QS BluetoothTile proiritizes connected device
When there is a connected device and the
Bluetooth Cacheded Device list is too large
it doesn't guarantee a connected device
will show up at all in the Bluetooth
Quick Settings Tile. This code modifies
the tile to match Wifi witch guarantees
connected devices be at the top of the
Bluetooth Tile list.

Bug: 37438599
Test: Manually pair multiple deivces to create
a large list and change connected states.

Change-Id: I93caa43743e1d63f4eab6b8bc636e8e8b2b40ec4

Former-commit-id: 6f991c296ad7469fcff58297b1e2632d787c24cc