DO NOT MERGE ANYWHERE: Move content from verison 123 to 122, and skip verison 123.
HFP profile config landed late in Emerald development, and was committed
as version 122 there.
When it merged to master, there were conflicts since 122 was already
taken, so it was resolved as version 123.
We didn't catch the subtle bug then:
This means Emerald devices will reset HFP profile config in error once
they upgrade to Feldspar.
Correct this by merging 122, and making 123 a no-op.

Bug: 37338202
Change-Id: Ibf4c824d27300641fd1bb15f5884baf056bd8d7e

Former-commit-id: d5f1a555f2fff12df95fc623fa3d9af623869bcb