Do not pause resumed activity in resumeTopActivityUncheckedLocked.
A number of operations occur between setting the resumed activity and
notifying the activity to resume. One of these steps is updating the
configuration, which can cause the device to pause the resumed
activity. Since the activity has not been told to resume, this leads
to a pause before resume.

This changelist addresses the issue by checking whether we are in the
middle of resuming the top activity before attempting to stop the
currently set resumed activity.

Change-Id: I3b5b61de6b1b5a35b7773912b4e63d6c69096c04
Fixes: 35637717
Test: bit FrameworksServicesTests:com.android.server.am.ActivityStackTests#testNoPauseDuringResumeTopActivity

Former-commit-id: 5daa3121a92f5e3124b8db469aeddea269a4c1df