Update fullscreen user switcher on the lock screen.
The older implementation didn't scale well with an
increasing number of users, it just kept making the
user pods smaller and smaller. Replace the custom
GridView with a custom ViewPager with a paging
indicator.

Also remove all the text from the screen since we
don't need it anymore.

More changes to follow.

Bug: 36454400
Test: Created more users than can fit on one screen,
      booted up and switched between the users.
Change-Id: I6de323fbaa6bc4cefc2bdb73a4c68c4f092a08a6

Former-commit-id: 5b3278f6147311d67f857c4339fd67dc44b6ff25