Always create display content with system identity
If an app tried to use a virtual display that it had just created
and add a window to it before it was registered in WM, WM would try
to create a DisplayContent instance. This would be executed with
app's calling identity and fail permission check.
This CL ensures that we always clear the calling identity before
creating display content.

Bug: 37422998
Test: android.display.cts.VirtualDisplayTest#testPrivateVirtualDisplay
Change-Id: I442cca65055886b384a28eeefcc35f2a36e482d0

Former-commit-id: 325cac4382259fd9b1f6b8e14840fe27d2d739e9