Fix potential NPE in Trampoline.requestBackup().
Before it was returning null in case the service was not initialized,
but as the method signature has int as the return type, this was boxed
and unboxed which would cause NPE.

Now it will return BackupManager.ERROR_BACKUP_NOT_ALLOWED instead.

This doesn't include fix for the test, because it's going to be
cherry-picked into oc-dev, which doesn't have the tests.

Test: manually ran unit tests

Bug: 37535747

Change-Id: I186cf812a4ec67be943b5c4334c82e9af5320b7a

Former-commit-id: 65090c6b47e8ae5e89fac4e0f664022d10de27e5