Tweak initialization order of LinearLayout's dividers
In ag/969570 we've unified the treatment of divider logic
in LinearLayout. However, setDividerDrawable logic that was
calling setWillNotDraw added condition that the divider
position is not NONE. That made the initial flow effectively
not calling setWillNotDraw(false).

Changing the order to call setDividerDrawable() at the end
invokes the right logic.

Test:  adb shell am instrument -e class 'android.widget.cts.LinearLayoutTest'  -w 'android.widget.cts/android.support.test.runner.AndroidJUnitRunner'
Bug: 37698004
Change-Id: I8d95770b9f48277c402708fbc5279ba2ec56719f

Former-commit-id: 3218a4e4f548a26e5783d3a56d150f07deeca448