Fix app transition logging when switching quickly
If an app was set to invisible before the transition started,
which happens when quickly open an app and then closing again
before the animation is starting (only possible if starting
window is slightly delayed), we never receive onWindowsDrawn so we
need to cancel the transition in this case.

Also fixes an issue where the wrong activity was reported as
starting to us in case reusedActivity != null but we'll only reuse
it for the task and not the actual activity.

Test: Open Calendar, press home immediately, open another app, make
sure all logged events and times are correct.
Fixes: 37538546

Change-Id: I5819bd97964ab2dd5cb40b2f4e3bc5a9ac348152

Former-commit-id: cdfc04e7b1924884fdb641a4f1adc07847e6ba36