Do not remove task until all activities have been destroyed.
Previously we were removing a task from its stack when the last
activity was removed from history or when there was only task
overlays present. This can lead to the situation where there is only
overlay activities present, meaning that we'll try to remove the
task again when those activities are removed from history.

This changelist addresses this issue by only removing the activity
when the last task is removed.

Change-Id: I2cc74b439934444d0694cd0d8bb611d4f9a6f5b3
Fixes: 37752430
Test: cts/hostsidetests/services/activityandwindowmanager/util/run-test CtsServicesHostTestCases android.server.cts.ActivityManagerPinnedStackTests#testFinishPipActivityWithTaskOverlay

Former-commit-id: 1d93058432445e2d63e7e1a467207fcc187c90a1