Statically include android.test.runner classes in legacy-android-test
In preparation for removing junit and android.test classes from
the Android API this change adds the android.test classes that are in
the android.test.runner library into legacy-android-test so that they
can be statically included into the application.

The junit.... classes from android.test.runner can already be
statically included from the junit library.

Bug: 30188076
Test: make checkbuild
Change-Id: Id8d32da9e46f0d9142b60ef28998b40e8d237b32

Former-commit-id: 4e0c19048d98f3057e5f4d20a2afd8a403087c30