WifiManagerTest: alt way to check callback
We are seeing failures on automated tests.  This is occuring when the
thread id is retrieved from the TestLooper.  Since we do not have
control over exactly which thread is used in the TestLooper for these
environments, this CL changes the checks to make sure the the callback
is called, and also that we never attempt to get the main looper (the
behavior when the provided handler is null).

Bug: 37828843
Test:  frameworks/base/wifi/tests/runtests.sh
Change-Id: I853c423b5f383225c250357d638c4cfd2c098edc

Former-commit-id: 695cea5929373a6eeda40578494cca214f1d83d2