TrustManagerService: Fix service binding leak
Fixes an issue where TrustAgentWrapper was created too early.
If we later decide to not actually start that agent, it is already
bound to, and then we just drop the TrustAgentWrapper class, never
unbinding.

In addition, this can also create a rebind loop where we are constantly
trying to rebind to the trust agent.

Bug: 37853858
Test: force-lock, observe that trust agents are properly unbound and not constantly trying to rebind
Change-Id: I4d383f9cb499874803a7f024915baf85b3682631

Former-commit-id: b884bb28b1416f8d0d93403eeac11af74b9f9968