Exit transient mode as soon as status bar window is fullscreen
When bouncer was popping up this caused us to call
PWM.adjustSystemUiVisibility from SystemUI. Since the transient
flag got cleared in that case we also hided the status bar, as we
assumed we are still in transient mode.

Instead, we exit transient mode as soon as the window is
fullscreen, which is quite robust.

This is not a perfect fix but given that the transient bar logic
is so quirky that needs to be rewritten at some point this fix
is good enough for now.

Test: Open camera, swipe down to show transient bars, click on the
Lock icon in the camera activity.
Fixes: 36333950

Change-Id: I6939c26f283d7d791c22d5891b62a705f62c116f

Former-commit-id: 2dc47c1dd55818a0bd68847215adccfe39d6dd88