Don't bring up backup agents in secondary app processes
When an app needs to be launched to make its agent available, we
were taking action in response to the first 'attach' operation we got
from any process running as the right uid.  This turns out to be fraught;
it's possible that some other secondary-process component launch was
racing with the agent bringup request, such that the attach callback
arrives from that secondary app process instead of the main process.
*That,* in turn, can be fatal if it turns out to be an isolated
process or similarly incapable host environment for the agent.

So now we make sure it's actually the process we're waiting for
when deciding to instantiate the agent there.

Bug 38123540
Test: manual

Change-Id: I01648027d104cc93fe0ec9227ebab85ad3d83c88

Former-commit-id: b73435cdedb03bbe03ccb3de3a51313d06ede09c