Always remove orphaned starting windows, no matter what the state
We orphaned a starting window because of the following sequence:
- Task was moved to front, thus starting window was added.
- activityPaused on Launcher was called
- In the meantime, any other activity was launched.
- When activityPause was completed, we didn't set the visibility
of the original task *not* to true. Thus, we didn't have any
signal to remove the starting window.

Note that this was more easy to repro with trampoline activities,
but in theory can also happen without trampoline activities.

We already have a way to remove orphaned starting windows so
let's reuse them independantly of the activity state.

Test: Hot launch calendar, immediately press home again. Make sure
not starting window is leaked.
Fixes: 37888853

Change-Id: Ide72c55e66a15b44a51380208bbb29dda00f7694

Former-commit-id: cb956050d5e4f7450dc8d6b362eff02455ebe264