Hiding text from obscured text fields.
My change to remove "speak passwords" set things up as if
the user had permanently enabled speaking of passwords.
That mode seems to go too far, though. It would speak the
password content even if it wasn't shown on the screen.

This change reverts a small part of ag/1900438 to avoid
populating text changed accessibility events with the plain
text of a view that is concealing that text from users who
are looking at the screen.

I'm tempted to populate the event instead with the
transformed text, but I'm concerned that corner cases may
exist where the transformed text is not valid in the
beforeTextChanged callback. This change seems safer, as it
is just going part way back to the pre-O behavior.

Bug: 38172502
Test: Verified that letters in the WiFi password are no
longer read out when they are deleted.

Change-Id: I1a000ee7f87c334a0d38248ce9ad120f43d12256

Former-commit-id: 52b9dbc67760307cf3a63bb2cde2f89233bc14ad