Lock out ResolverDrawerLayout dismissals during animations
Due to views being offset during the ChooserActivity animating in
direct share targets, it was possible to touch down on the sheet
attempting to tap, but have it be interpreted as a fling downward as
the relative touch position within the parent was moving. Lock the
dismiss gesture during animations to prevent this.

Bug: 36032762
Test: manual
Change-Id: I18b6f03148cbf402efc23e873262927e412ea7b7

Former-commit-id: 08adbfed3a3c291a3c9f1cb1af77dd5930a17dd1