Add a method that allows the shade to be toggled.
This method allows the shade to be dismissed when it's down and vice
versa.

An alternative is to expose a method that returns a boolean indicating
whether or not the shade is down and call the show/hide methods
appropriately. However, choose not to go with this approach since
IStatusBar.aidl is oneway.

Test: booted up on phone and AAE headunit
Bug: 35930275
Change-Id: I5b8cadc1793ff2e21f242e866c01f57de20fbea6

Former-commit-id: 9ad00e090d2a6d979dbc6e7eae056e1655c8ef2b