Always add the instant app installer
Even if we filter out an instant app because the caller doesn't
have access to it, always make sure to add the installer when
we have the possibility of an instant app.

NOTE: This does not have CTS tests since it's impossible to test
without involving the resolver. CTS tests will be added in MR1+
when we can make changes to the APIs and add the necessary methods
to mock out the resolver service.

Change-Id: I92c88eb04520c0c5e0c58db6d668ed0a0e552a35
Fixes: 38268052
Test: cts-tradefed run commandAndExit cts-dev -m CtsAppSecurityHostTestCases -t android.appsecurity.cts.EphemeralTest

Former-commit-id: 22d8bb2ca310a929803be4c4ab99c44bc43d1a93