Reset hidden state of AppToken that was cleared due to starting window
If a starting window is tranferred to a token from another token that is
visible, we also make the receiving token visible, however we don't
reset the start when the starting window is tranferred out again when no
other windows were added to the token. We now do.

Fixes: 38317227
Test: Launch Audible app music player and verify the LibraryActivity
token isn't in the visible state.

Change-Id: I0186c79aa24d64362f742229d60c75eb0c1a3f97

Former-commit-id: 6c459219b0bcd6181a45a4bb3ae140264fc9e146