Wait on activity brought to front for trampoline activities.
We were previously waiting on a trampoline activity to become
visible when an existing activity (when reused via intent match) was
brought to front. This will never resolve as the trampoline activity
does not become visible.

This CL identifies this situation and properly associates the wait
with the reused activity.

Change-Id: I55f47766e8cdf92e31b4ddd97864cb0f5f3b751a
Fixes: 38030545
Test: cts-tradefed run cts -m CtsServicesHostTestCases -t  android.server.cts.ActivityManagerAmStartOptionsTests#testDashW_Indirect

Former-commit-id: 89cd19a4c4d274438ec63f79aae091c37e801f18