docs: Clarifying code snippet for View.autoFill()
Per bug, updating code snippet to clarify that it's an example of an
implementation of the method. Oscar & Felipe, please check that I did
it right!

Also fixed a couple of HTML syntax errors while I had the file open
(badly formatted escape-characters that Chrome figured out, but other
browsers might choke on) and a spelling error.

Revised Javadoc is staged to:

http://go/dac-stage/reference/android/view/View.html#autofill(android.view.autofill.AutofillValue)

Test: make ds-docs
Bug: 38347106
Change-Id: I587a66c53fd5ebeeb6108529723d2d7a74c61cf7

Former-commit-id: 05ceb4e0a6e5a986ddf665e9b52123a8c533d6b4