Best effort dismissal of the login screen.
This change adds in two mechanisms to dismiss the login
screen that shows up on boot. The first is a countdown
timer and the second is a start driving button. Both of
them behave in the same way that once the timer runs out
or if the button is clicked, the system will automatically
try to select a user and get the system running.

Right now the implementation of selecting a user is a bit
trivial, it just tries to drop the keyguard on the currently
running user.

Bug: 36454400
Test: 1. Booted mojave, watched progress bar make progress
         and automatically log in user.
      2. Booted mojave, clicked the start driving button and
         watch the keyguard dismiss itself.
Change-Id: If22f4c3c7e21aadf2951b9c0cd8b97a4b474b349

Former-commit-id: 97ff85838e21c18fbfc2e509bcc92049716ab97c